Moved UILablel  highlight extension to separate file
